backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
scikit-learn
python-multipart
httpx
